Tests and Fixes for mobile

+++ SimpliPassMobile/SimpliPassMobile/Models/SectionModel.cs
@@ -7,6 +7,6 @@
     {
         public string Name { get; set; }
         public double Rating { get; set; }
-        public double Count { get; set; }
+        public int Count { get; set; }
     }
 }
